% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{coef_ci}
\alias{coef_ci}
\title{'Finds the CI for regression coefficients'}
\usage{
coef_ci(data)
}
\arguments{
\item{data}{with y = input$response, and x = input$subsetted_columns}
}
\description{
confidence interval for betas, multivariate regression coefficients
}
\examples{
y= flights$time
x = flights[,c(1:3)]
data  = data.frame(y,x)
coef_ci(data)
}
