% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{pred_ci}
\alias{pred_ci}
\title{'Finds the CI for new prediction'}
\usage{
pred_ci(data, newdata)
}
\arguments{
\item{data}{same data used as input for coef_ci}

\item{newdata}{a dataframe with test input values for the explanatory columns}
}
\description{
confidence interval for new prediction
}
\examples{
newdata = data.frame(dep_delay = c(2,3), arr_delay = c(13,-15), distance = c(450,500))
pred_ci(data,newdata)
}
